requirements.txt
streamlit
python-dotenv
httpx
llama-index
